convert int to long long

+++ src/dem_disc_main.c
@@ -9,10 +9,10 @@ int main(int argc, char **argv)
         printf("Usage: %s [file]\n", argv[0]);
         return 1;
     }
-    int n, d;
+    long long n, d;
     double *points_store = read_points_from_file(argv[1], &d, &n);
     double **points = malloc(n * sizeof(double *));
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         points[i] = points_store + i * d;
     }
     double discrepancy = oydiscr(points, d, n);
